@@ -15,14 +15,14 @@
   ],
   "minimum-stability": "dev",
   "require": {
-    "php": "^7.2",
-    "zendframework/zend-mvc": "^3.0.1",
-    "zendframework/zend-mvc-form": "^1.0.0",
-    "solarium/solarium": "^4.1",
+    "php": "^7.4",
+    "laminas/laminas-mvc": "^3.1",
+    "laminas/laminas-mvc-form": "^1.0.0",
+    "solarium/solarium": "5.*",
     "phpoffice/phppresentation": "^0.9"
   },
   "require-dev": {
-    "iteaoffice/testing": "^1.2"
+    "iteaoffice/testing": "^2.1"
   },
   "autoload": {
     "psr-4": {
@@ -31,13 +31,13 @@
   },
   "autoload-dev": {
     "psr-4": {
-      "ProjectTest\\": "test/"
+      "SearchTest\\": "test/"
     }
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "1.0.x-dev",
-      "dev-develop": "1.1.x-dev"
+      "dev-master": "2.0-DEV",
+      "dev-develop": "2.1-DEV"
     }
   }
 }
