@@ -1,4 +1,5 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
@@ -7,7 +8,7 @@
  * @category    Search
  *
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2018 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license     https://itea3.org/license.txt proprietary
  *
  * @link        http://github.com/iteaoffice/main for the canonical source repository
@@ -20,13 +21,15 @@ namespace Search\Service;
 use Affiliation\Service\AffiliationService;
 use Calendar\Service\CalendarService;
 use Contact\Service\ContactService;
-use Contact\Service\ProfileService;
 use Event\Service\RegistrationService;
 use General\Service\CountryService;
+use InvalidArgumentException;
 use Invoice\Service\InvoiceService;
+use News\Service\BlogService;
 use News\Service\NewsService;
 use Organisation\Service\OrganisationService;
 use Press\Service\PressService;
+use Project\Service\ActionService;
 use Project\Service\DescriptionService;
 use Project\Service\IdeaService;
 use Project\Service\ProjectService;
@@ -43,74 +46,25 @@ use Publication\Service\PublicationService;
  */
 class ConsoleService
 {
-    /**
-     * @var ContactService
-     */
-    private $contactService;
-    /**
-     * @var IdeaService
-     */
-    private $ideaService;
-    /**
-     * @var DescriptionService
-     */
-    private $descriptionService;
-    /**
-     * @var ProjectService
-     */
-    private $projectService;
-    /**
-     * @var VersionService
-     */
-    private $versionService;
-    /**
-     * @var VersionDocumentService
-     */
-    private $versionDocumentService;
-    /**
-     * @var WorkpackageService
-     */
-    private $workpackageService;
-    /**
-     * @var ResultService
-     */
-    private $resultService;
-    /**
-     * @var PublicationService
-     */
-    private $publicationService;
-    /**
-     * @var InvoiceService
-     */
-    private $invoiceService;
-    /**
-     * @var AffiliationService
-     */
-    private $affiliationService;
-    /**
-     * @var RegistrationService
-     */
-    private $registrationService;
-    /**
-     * @var CalendarService
-     */
-    private $calendarService;
-    /**
-     * @var NewsService
-     */
-    private $newsService;
-    /**
-     * @var PressService
-     */
-    private $pressService;
-    /**
-     * @var OrganisationService
-     */
-    private $organisationService;
-    /**
-     * @var CountryService
-     */
-    private $countryService;
+    private ContactService $contactService;
+    private IdeaService $ideaService;
+    private DescriptionService $descriptionService;
+    private ProjectService $projectService;
+    private VersionService $versionService;
+    private VersionDocumentService $versionDocumentService;
+    private WorkpackageService $workpackageService;
+    private ResultService $resultService;
+    private ActionService $actionService;
+    private PublicationService $publicationService;
+    private InvoiceService $invoiceService;
+    private AffiliationService $affiliationService;
+    private RegistrationService $registrationService;
+    private CalendarService $calendarService;
+    private NewsService $newsService;
+    private BlogService $blogService;
+    private PressService $pressService;
+    private OrganisationService $organisationService;
+    private CountryService $countryService;
 
     public function __construct(
         ContactService $contactService,
@@ -121,12 +75,14 @@ class ConsoleService
         VersionDocumentService $versionDocumentService,
         WorkpackageService $WorkpackageService,
         ResultService $resultService,
+        ActionService $actionService,
         PublicationService $publicationService,
         InvoiceService $invoiceService,
         AffiliationService $affiliationService,
         RegistrationService $registrationService,
         CalendarService $calendarService,
         NewsService $newsService,
+        BlogService $blogService,
         PressService $pressService,
         OrganisationService $organisationService,
         CountryService $countryService
@@ -139,12 +95,14 @@ class ConsoleService
         $this->versionDocumentService = $versionDocumentService;
         $this->workpackageService = $WorkpackageService;
         $this->resultService = $resultService;
+        $this->actionService = $actionService;
         $this->publicationService = $publicationService;
         $this->invoiceService = $invoiceService;
         $this->affiliationService = $affiliationService;
         $this->registrationService = $registrationService;
         $this->calendarService = $calendarService;
         $this->newsService = $newsService;
+        $this->blogService = $blogService;
         $this->pressService = $pressService;
         $this->organisationService = $organisationService;
         $this->countryService = $countryService;
@@ -183,20 +141,23 @@ class ConsoleService
             case 'result':
                 $this->resultService->updateCollectionInSearchEngine($clearIndex);
                 break;
+            case 'action':
+                $this->actionService->updateCollectionInSearchEngine($clearIndex);
+                break;
             case 'publication':
                 $this->publicationService->updateCollectionInSearchEngine($clearIndex);
                 break;
             case 'invoice':
                 $this->invoiceService->updateCollectionInSearchEngine($clearIndex);
                 break;
-            case 'affiliation':
-                $this->affiliationService->updateCollectionInSearchEngine($clearIndex);
-                break;
             case 'calendar':
                 $this->calendarService->updateCollectionInSearchEngine($clearIndex);
                 break;
             case 'news':
                 $this->newsService->updateCollectionInSearchEngine($clearIndex);
+                break;
+            case 'blog':
+                $this->blogService->updateCollectionInSearchEngine($clearIndex);
                 break;
             case 'press':
                 $this->pressService->updateCollectionInSearchEngine($clearIndex);
@@ -215,25 +176,21 @@ class ConsoleService
                 $this->descriptionService->updateCollectionInSearchEngine($clearIndex);
                 $this->projectService->updateCollectionInSearchEngine($clearIndex);
                 $this->versionService->updateCollectionInSearchEngine($clearIndex);
+                $this->actionService->updateCollectionInSearchEngine($clearIndex);
                 $this->versionDocumentService->updateCollectionInSearchEngine($clearIndex);
                 $this->workpackageService->updateCollectionInSearchEngine($clearIndex);
                 $this->resultService->updateCollectionInSearchEngine($clearIndex);
                 $this->publicationService->updateCollectionInSearchEngine($clearIndex);
                 $this->invoiceService->updateCollectionInSearchEngine($clearIndex);
-                $this->affiliationService->updateCollectionInSearchEngine($clearIndex);
                 $this->calendarService->updateCollectionInSearchEngine($clearIndex);
                 $this->newsService->updateCollectionInSearchEngine($clearIndex);
+                $this->blogService->updateCollectionInSearchEngine($clearIndex);
                 $this->pressService->updateCollectionInSearchEngine($clearIndex);
                 $this->organisationService->updateCollectionInSearchEngine($clearIndex);
                 $this->countryService->updateCollectionInSearchEngine($clearIndex);
                 break;
             default:
-                throw new \InvalidArgumentException(sprintf("%s is incorrect", $index));
+                throw new InvalidArgumentException(sprintf('%s is incorrect', $index));
         }
-    }
-
-    public function updateRegistrationByMeeting(int $meetingId, bool $clearIndex = false): void
-    {
-        $this->registrationService->updateRegistrationByMeeting($meetingId, $clearIndex);
     }
 }
