@@ -5,7 +5,7 @@
  * @category  Search
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 use Search\Controller;
@@ -13,7 +13,7 @@ use Search\Controller;
 return [
     'router' => [
         'routes' => [
-            'search'   => [
+            'search' => [
                 'type'     => 'Literal',
                 'priority' => 1000,
                 'options'  => [
