@@ -1,4 +1,5 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
@@ -6,7 +7,7 @@
  *
  *
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2018 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license     https://itea3.org/license.txt proprietary
  *
 
@@ -16,15 +17,15 @@ declare(strict_types=1);
 
 namespace Search;
 
-use Zend\Console\Adapter\AdapterInterface;
-use Zend\ModuleManager\Feature;
+use Laminas\Console\Adapter\AdapterInterface;
+use Laminas\ModuleManager\Feature;
 
 /**
  * Class Module
  *
  * @package Search
  */
-class Module implements
+final class Module implements
     Feature\ConfigProviderInterface,
     Feature\ConsoleUsageProviderInterface
 {
@@ -38,25 +39,19 @@ class Module implements
         return [
             'Search management',
             // Describe available commands
-            'search reset <entity>'                => 'Reset the search index (wipe and rebuild index)',
+            'search reset <entity>' => 'Reset the search index (wipe and rebuild index)',
             [
                 '<entity>',
-                'Choose an entity to reset: project, version, roadmap, version-document, workpackage-document, result, impact-stream, publication, invoice, contact, profile, affiliation, registration, calendar, news, press, organisation, country all',
+                'Choose an entity to reset: project, version, roadmap, version-document, 
+                workpackage-document, result, impact-stream, action, publication, invoice, contact, 
+                profile, affiliation, registration, calendar, news, blog, press, organisation, country, all',
             ],
-            'search update <entity>'               => 'Update the search index',
+            'search update <entity>' => 'Update the search index',
             [
                 '<entity>',
-                'Choose an entity to update: project, version, roadmap, version-document, workpackage-document, result, impact-stream, publication, invoice, contact, profile, affiliation, registration, calendar, news, press, organisation, country all',
-            ],
-            'search reset-registration <meeting>'  => 'Reset the registrations of a meeting',
-            [
-                '<meeting>',
-                'Give the ID of the meeting',
-            ],
-            'search update-registration <meeting>' => 'Update the registrations of a meeting',
-            [
-                '<meeting>',
-                'Give the ID of the meeting',
+                'Choose an entity to update: project, version, roadmap, version-document, 
+                workpackage-document, result, impact-stream, action, publication, invoice, contact, 
+                profile, affiliation, registration, calendar, news, blog, press, organisation, country, all',
             ],
         ];
     }
